requirements.txt
fastapi
uvicorn
groq
sqlalchemy
python-dotenv
chromadb
requests
psycopg2-binary
pypdf2